refactor(docker): rewrite configuration system.
